chore: code clean up

+++ BetaSharp.Client/Rendering/Blocks/Renderers/TorchRenderer.cs
@@ -17,33 +17,33 @@ public class TorchRenderer : IBlockRenderer
 
         ctx.Tess.setColorOpaque_F(luminance, luminance, luminance);
 
-        double tiltAmount = 0.4F;
-        double horizontalOffset = 0.5D - tiltAmount;
-        double verticalOffset = 0.2F;
+        float tiltAmount = 0.4f;
+        float horizontalOffset = 0.5f - tiltAmount;
+        float verticalOffset = 0.2f;
 
         if (metadata == 1) // Attached to West wall (pointing East)
         {
             ctx.DrawTorch(block, new Vec3D(pos.x - horizontalOffset, pos.y + verticalOffset, pos.z),
-                -tiltAmount, 0.0D);
+                -tiltAmount, 0.0f);
         }
         else if (metadata == 2) // Attached to East wall (pointing West)
         {
             ctx.DrawTorch(block, new Vec3D(pos.x + horizontalOffset, pos.y + verticalOffset, pos.z),
-                tiltAmount, 0.0D);
+                tiltAmount, 0.0f);
         }
         else if (metadata == 3) // Attached to North wall (pointing South)
         {
-            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y + verticalOffset, pos.z - horizontalOffset), 0.0D,
+            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y + verticalOffset, pos.z - horizontalOffset), 0.0f,
                 -tiltAmount);
         }
         else if (metadata == 4) // Attached to South wall (pointing North)
         {
-            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y + verticalOffset, pos.z + horizontalOffset), 0.0D,
+            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y + verticalOffset, pos.z + horizontalOffset), 0.0f,
                 tiltAmount);
         }
         else // Standing on floor
         {
-            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y, pos.z), 0.0D, 0.0D);
+            ctx.DrawTorch(block, new Vec3D(pos.x, pos.y, pos.z), 0.0f, 0.0f);
         }
 
         return true;
